#339 fix: Fixed the DataTable -- it does not have any pagination now
